+++ examples/libxcb.rs
@@ -17,6 +17,7 @@ mod example {
         DisplayHandle, RawDisplayHandle, RawWindowHandle, WindowHandle, XcbDisplayHandle,
         XcbWindowHandle,
     };
+    use softbuffer::{Context, Pixel, Surface};
     use std::{env, num::NonZeroU32, ptr::NonNull};
     use x11rb::{
         connection::Connection,
@@ -26,8 +27,6 @@ mod example {
         },
         xcb_ffi::XCBConnection,
     };
-
-    const RED: u32 = 255 << 16;
 
     pub(crate) fn run() {
         // Create a new XCB connection
@@ -78,8 +77,8 @@ mod example {
             unsafe { DisplayHandle::borrow_raw(RawDisplayHandle::Xcb(display_handle)) };
         let window_handle =
             unsafe { WindowHandle::borrow_raw(RawWindowHandle::Xcb(window_handle)) };
-        let context = softbuffer::Context::new(display_handle).unwrap();
-        let mut surface = softbuffer::Surface::new(&context, window_handle).unwrap();
+        let context = Context::new(display_handle).unwrap();
+        let mut surface = Surface::new(&context, window_handle).unwrap();
 
         // Register an atom for closing the window.
         let wm_protocols_atom = conn
@@ -124,7 +123,7 @@ mod example {
                         )
                         .unwrap();
                     let mut buffer = surface.buffer_mut().unwrap();
-                    buffer.pixels().fill(RED);
+                    buffer.pixels().fill(Pixel::new_rgb(0xff, 0, 0));
                     buffer.present().unwrap();
                 }
                 Event::ConfigureNotify(configure_notify) => {